Fixed Animal test (now parameterized) 100% test coverage provided
